Magically bluetooth working better
I'm totally unsure if this is doing anything useful, but it seems
that bluetooth is more reliable by doing a scan for 5 seconds when problems
pop up. An example problem is that the UI text says "null - connecting". This
means some device has a null pointer name... but if I scan for 5 seconds this
seems to go away. Retrying bluetooth connect can help.