Add auto siren. Clean up toneGenerator
Still kind of a work in progress. Specifically what we want is that the
next 250ms after any excessive movement, the siren is on. Right now,
what may happen is that there's a bump at:
t=0, t=200ms, t=300ms. What will happen is the siren will go off at
250ms and back on at 300.. it should just stay on the whole time!